review: move test to Bun-specific file, restore async context on early return
- moved the regression test from node-http-backpressure.test.ts (Node-compat
  file) to node-http-uaf.test.ts which already holds Bun-specific subprocess
  crash regressions
- Bun__JSValue__call: restore the swapped async context before the new
  non-callable early return, and guard the debug auditCellFully with isCell()
  since the value may now be a primitive
